refactor: logo/ 디렉토리 icon/ 으로 이름 변경
