+++ .gitignore
@@ -8,7 +8,7 @@
 .LSOverride
 
 # Icon must end with two \r
-Icon
+
 
 
 # Thumbnails
