+++ apps/client/src/components/icon/Harmony.tsx
