+++ apps/client/src/components/icon/HarmonyWithText.tsx
