+++ apps/client/src/components/icon/index.ts
@@ -1,0 +1,5 @@
+import Github from '@/components/icon/GitHub';
+import HarmonyWithText from '@/components/icon/HarmonyWithText';
+import Harmony from '@/components/icon/Harmony';
+
+export { HarmonyWithText, Harmony, Github };
