--- apps/client/src/components/logo/index.ts
@@ -1,5 +1,0 @@
-import Github from '@/components/logo/GitHub';
-import HarmonyWithText from '@/components/logo/HarmonyWithText';
-import Harmony from '@/components/logo/Harmony';
-
-export { HarmonyWithText, Harmony, Github };
